move style related plot funcs
