+++ spiketools/tests/plts/test_style.py
@@ -1,0 +1,45 @@
+"""Tests for spiketools.plts.style."""
+
+from spiketools.plts.style import *
+
+###################################################################################################
+###################################################################################################
+
+def test_get_set_kwargs():
+
+    kwargs = {'title' : 'title', 'xlabel' : 'xlabel', 'lw' : 12}
+    out = get_set_kwargs(kwargs)
+    for arg in ['title', 'xlabel']:
+        assert arg in out
+        assert arg not in kwargs
+    for arg in ['lw']:
+        assert arg not in out
+        assert arg in kwargs
+
+def test_get_attr_kwargs():
+
+    kwargs = {'title_color' : 'red', 'xlabel' : 'xlabel', 'lw' : 12}
+    out = get_attr_kwargs(kwargs, 'title')
+    for arg in ['color']:
+        assert arg in out
+        assert 'title_' + arg not in kwargs
+    for arg in ['xlabel', 'lw']:
+        assert arg not in out
+        assert arg in kwargs
+
+def test_set_plt_kwargs():
+
+    @set_plt_kwargs
+    def example_plot():
+        plt.plot([1, 2], [3, 4])
+
+    xlim = (0, 2)
+    ylim = (0, 4)
+    title = 'Test Title'
+
+    example_plot(xlim=xlim, ylim=ylim)
+    assert plt.gca().get_xlim() == xlim
+    assert plt.gca().get_ylim() == ylim
+
+    example_plot(title=title)
+    assert plt.gca().get_title() == title
